requirements.txt
pandas
numpy
pytorch
transformers
yargy